chore: move ParseEnvVariablesFromFile to skaffold pkg util package
