--- pkg/skaffold/verify/util/util.go
@@ -1,36 +1,0 @@
-/*
-Copyright 2023 The Skaffold Authors
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package util
-
-import (
-	"os"
-
-	"github.com/joho/godotenv"
-)
-
-// ParseEnvVariablesFromFile takes an env file at the given filepath `fp` and parse the env values into a map
-func ParseEnvVariablesFromFile(fp string) (map[string]string, error) {
-	f, err := os.Open(fp)
-	if err != nil {
-		return nil, err
-	}
-	envMap, err := godotenv.Parse(f)
-	if err != nil {
-		return nil, err
-	}
-	return envMap, nil
-}
